Added initial countdown
Slightly modified the structure of the Engine to allow for an initial
countdown. Now the match is not immediately added to the list of games
the Engine holds. Instead the Hub is immediately sending to the clients
the init message, and then sleeps for 3 seconds, the same amount of time
that the client will need to perform the countdown (of course we are not
taking into consideration network lag), and finally starts simulating